requirements.txt
Flask==3.1.2
Flask_Cors==5.0.0
flask_limiter==3.12
moviepy==2.0.0
Pillow==9.3.0
Pillow==12.0.0
python-dotenv==1.2.1
Requests==2.32.5
user_agents==2.2.0
Werkzeug==2.2.3